refactor: remove facilities page and route directory
