+++ bciers/apps/administration/app/bceidbusiness/industry_user/operations/[operationId]/facilities/page.tsx
@@ -1,8 +1,8 @@
 // 🚩 flagging that for shared routes between roles, "Page" code is a component for code maintainability
 import { FacilitiesSearchParams } from "@/administration/app/components/facilities/types";
 import FacilitiesPage, {
-  ExternalFacilitiesLayout,
-} from "@/administration/app/components/routes/facilities/Page";
+  ExternalUserFacilitiesLayout,
+} from "@/administration/app/components/facilities/Facilities";
 import { UUID } from "crypto";
 
 export default async function Page({
@@ -13,8 +13,8 @@ export default async function Page({
   searchParams: FacilitiesSearchParams;
 }>) {
   return (
-    <ExternalFacilitiesLayout>
+    <ExternalUserFacilitiesLayout>
       <FacilitiesPage operationId={operationId} searchParams={searchParams} />
-    </ExternalFacilitiesLayout>
+    </ExternalUserFacilitiesLayout>
   );
 }
