+++ bciers/apps/administration/app/bceidbusiness/industry_user_admin/operations/page.tsx
@@ -1,7 +1,7 @@
 // 🚩 flagging that for shared routes between roles, "Page" code is a component for code maintainability
 import { OperationsSearchParams } from "@/app/components/operations/types";
 import OperationsPage, {
-  ExternalOperationsLayout,
+  ExternalUserOperationsLayout,
 } from "@/administration/app/components/operations/Operations";
 
 export default async function Page({
@@ -10,8 +10,8 @@ export default async function Page({
   searchParams: OperationsSearchParams;
 }) {
   return (
-    <ExternalOperationsLayout>
+    <ExternalUserOperationsLayout>
       <OperationsPage searchParams={searchParams} isInternalUser={false} />
-    </ExternalOperationsLayout>
+    </ExternalUserOperationsLayout>
   );
 }
